WinAnalytics moved to WinMenuScript
